Next.js 테스트 배포, Github -> EC2 CI/CD 적용

+++ scripts/after-install.sh
@@ -1,4 +1,18 @@
-  #!/bin/bash
-  cd /var/frontend/
+#!/bin/bash
+export NVM_DIR="$([ -z "${XDG_CONFIG_HOME-}" ] && printf %s "${HOME}/.nvm" || printf %s "${XDG_CONFIG_HOME}/nvm")"
+[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh" # This loads nvm
+source /home/ec2-user/.bash_profile
+
+cd /var/frontend/
+  
+# Check if the node_modules directory exists
+if [ -d "node_modules/" ]; then
+  # If it exists, remove it
+  echo "node_modules/ directory exists. Removing it..."
   rm -r node_modules/
-  pnpm install
+else
+  # If it doesn't exist, print a message
+  echo "node_modules/ directory does not exist. Nothing to do."
+fi 
+
+pnpm install
